2-Langchain/requirements.txt
langchain
langchain_community
python-dotenv
openai
langchian_groq
fastapi
uvicorn
langserve
sse_starlette